requirements.txt
openai
python-dotenv 
psycopg2-binary 
redis 
rq